requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
twilio
sendgrid
python-dotenv
python-multipart
pydantic[email]